게시글 볼 때 member의 nickname이 writer에 표시(버그 수정)

+++ src/main/java/com/example/prj1fe20231109mj/controller/MemberController.java
@@ -63,20 +63,44 @@ public class MemberController {
         return service.list();
     }
 
+
+
+    // 회원 정보 보기
     @GetMapping
-    public ResponseEntity<Member> view(String id) {
+    public ResponseEntity<Member> view(String id,
+                                       @SessionAttribute(value = "login", required = false) Member login) {
         // TODO : 로그인 했는 지? -> 안했으면 401
         // TODO : 자기 정보인지? -> 아니면 403
+
+        if (login == null) {
+
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if (!service.hasAccess(id, login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         Member member = service.getMember(id);
 
         return ResponseEntity.ok(member);
     }
 
+
+    // 탈퇴하려고 할 때
     @DeleteMapping
-    public ResponseEntity delete(String id) {
+    public ResponseEntity delete(String id,
+                                 @SessionAttribute(value = "login", required = false) Member login) {
         // TODO : 로그인 했는 지? -> 안했으면 401
         // TODO : 자기 정보인지? -> 아니면 403
+
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+        if (!service.hasAccess(id, login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
 
         if (service.deleteMember(id)) {
             return ResponseEntity.ok().build();
@@ -87,8 +111,17 @@ public class MemberController {
     }
 
     @PutMapping("edit")
-    public ResponseEntity edit(@RequestBody Member member) {
+    public ResponseEntity edit(@RequestBody Member member,
+                               @SessionAttribute(value = "login",required = false) Member login) {
         // TODO: 로그인 했는지? 자기정보인지?
+
+        if (login == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+        if (!service.hasAccess(member.getId(), login)){
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 401
+        }
 
         if (service.update(member)) {
             return ResponseEntity.ok().build();
